# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movie Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stickers/upload-url:
    post:
      operationId: getStickerUploadUrl
      tags: [stickers]
      summary: Get a signed URL for uploading a sticker image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetUploadUrlBody"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /stickers:
    get:
      operationId: listGlobalStickers
      tags: [stickers]
      summary: List all global stickers
      responses:
        "200":
          description: Global stickers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalStickersResponse"
    post:
      operationId: createGlobalSticker
      tags: [stickers]
      summary: Create a global sticker (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerBody"
      responses:
        "201":
          description: Sticker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"
  /stickers/{stickerId}:
    delete:
      operationId: deleteGlobalSticker
      tags: [stickers]
      summary: Delete a global sticker (super admin only)
      parameters:
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sticker deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteStickerResponse"
  /groups/{groupId}/stickers:
    get:
      operationId: listGroupStickers
      tags: [stickers]
      summary: List stickers available for a group (global + group-specific)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stickers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupStickersResponse"
    post:
      operationId: createGroupSticker
      tags: [stickers]
      summary: Create a group-specific sticker (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerBody"
      responses:
        "201":
          description: Sticker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"
  /groups/{groupId}/stickers/{stickerId}:
    delete:
      operationId: deleteGroupSticker
      tags: [stickers]
      summary: Delete a group sticker (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sticker deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteStickerResponse"

  # Reaction endpoints
...
components:
  schemas:
...
    Sticker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
        isGlobal:
          type: boolean
        groupId:
          type: integer
          nullable: true
        createdBy:
          type: integer
      required:
        - id
        - name
        - imageUrl
        - isGlobal
    GetUploadUrlBody:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
        groupId:
          type: integer
          nullable: true
      required:
        - filename
        - contentType
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        objectName:
          type: string
        publicUrl:
          type: string
      required:
        - uploadUrl
        - objectName
        - publicUrl
    CreateStickerBody:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - imageUrl
    GlobalStickersResponse:
      type: object
      properties:
        stickers:
          type: array
          items:
            $ref: "#/components/schemas/Sticker"
      required:
        - stickers
    GroupStickersResponse:
      type: object
      properties:
        globalStickers:
          type: array
          items:
            $ref: "#/components/schemas/Sticker"
        groupStickers:
          type: array
          items:
            $ref: "#/components/schemas/Sticker"
        stickers:
          type: array
          items:
            $ref: "#/components/schemas/Sticker"
      required:
        - globalStickers
        - groupStickers
        - stickers
    DeleteStickerResponse:
      type: object
      properties:
        message:
          type: string
        reactionsRemoved:
          type: integer
      required:
        - message
        - reactionsRemoved

    # Reaction schemas
...