# lib/api-zod/src/generated/types/sticker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface Sticker {
  id: number;
  name: string;
  imageUrl: string;
  isGlobal: boolean;
  groupId?: number | null;
  createdBy?: number;
}