# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie Club API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlResponse {
  uploadUrl: string;
  objectName: string;
  publicUrl: string;
}